+++ configure.py
@@ -34,6 +34,11 @@ GPU_ARCHS = frozenset(
 KNOWN_ARCHS = {"CPU": CPU_ARCHS, "GPU": GPU_ARCHS}
 DEFAULT_ARCHS = {"CPU": "BDW", "GPU": "Pascal60"}
 
+_GPU_FLAG_DEPRECATION_MESSAGE = (
+    "The -gpu flag is deprecated. Using it will raise an error in a future release. "
+    "Please explicitly request a GPU architecture via the -arch argument."
+)
+
 
 def _add_parser_args(parser):
     parser.add_argument(
@@ -41,11 +46,11 @@ def _add_parser_args(parser):
     )
 
     parser.add_argument("-mhd", action="store_true", help="enable MHD")
+
     parser.add_argument(
         "-gpu",
-        dest="use_gpu",
         action="store_true",
-        help="enable KOKKOS+CUDA",
+        help=_GPU_FLAG_DEPRECATION_MESSAGE,
     )
 
     parser.add_argument("-cxx", help="override default compiler")
@@ -54,21 +59,28 @@ def _add_parser_args(parser):
         "-arch",
         nargs="+",
         dest="archs",
-        default=[DEFAULT_ARCHS["CPU"], DEFAULT_ARCHS["GPU"]],
         choices=CPU_ARCHS.union(GPU_ARCHS),
         help="target Kokkos architecture (accepts up to one CPU and up to one GPU archs)",
     )
     parser.add_argument(
         "-openmp",
-        help="enable OpenMP parallelism (not available with -gpu)",
+        help="enable OpenMP parallelism (not available with GPU architectures)",
         action="store_true",
     )
     parser.add_argument("-mpi", action="store_true", help="enable MPI parallelism")
 
 
+def is_gpu_requested(requested_archs):
+    if requested_archs is None:
+        return False
+    return any((a in GPU_ARCHS for a in requested_archs))
+
+
 def parse_archs(requested_archs):
     # parse architectures:
     # at most 2 can be specified by the user, but only one for each arch type (CPU, GPU)
+    if requested_archs is None:
+        return DEFAULT_ARCHS
     if len(requested_archs) > 2:
         raise ValueError(
             "Error: received more than two architectures ({}).".format(
@@ -170,7 +182,6 @@ def _write_makefile(
 
 def _get_report(
     archs,
-    use_gpu,
     openmp,
     mpi,
     mhd,
@@ -191,7 +202,7 @@ def _get_report(
     ]
 
     selected_archs = parse_archs(archs)
-    arch_type = "GPU" if use_gpu else "CPU"
+    arch_type = "GPU" if is_gpu_requested(archs) else "CPU"
     report_lines += [
         "Execution target: {}".format(arch_type),
         "Target architecture: {}".format(selected_archs[arch_type]),
@@ -227,13 +238,23 @@ def main(argv=None):
         print(err, file=sys.stderr)
         return 1
 
-    if args.openmp and args.use_gpu:
-        print("Warning: with -gpu, -openmp flag is ignored.", file=sys.stderr)
+    use_gpu = is_gpu_requested(args.archs)
+    if args.gpu:
+        print("Warning: " + _GPU_FLAG_DEPRECATION_MESSAGE, file=sys.stderr)
+        if not use_gpu:
+            print(
+                "Warning: -gpu flag was received, but no GPU architecture was specified. "
+                "Defaulting to %s" % selected_archs["GPU"],
+                file=sys.stderr,
+            )
+
+    if args.openmp and use_gpu:
+        print("Warning: with a GPU arch, -openmp flag is ignored.", file=sys.stderr)
         args.openmp = False
 
     makefile_options = _get_makefile_options(
         archs=selected_archs,
-        use_gpu=args.use_gpu,
+        use_gpu=use_gpu,
         cxx=args.cxx,
         openmp=args.openmp,
         mpi=args.mpi,
@@ -248,7 +269,6 @@ def main(argv=None):
 
     report = _get_report(
         archs=args.archs,
-        use_gpu=args.use_gpu,
         openmp=args.openmp,
         mpi=args.mpi,
         mhd=args.mhd,